Added parametrized template
The X17 is currently a template computed with a gaussian PDF. It is
treated as any other histogram. The total statistics of X17 is hard
coded and does not need to be too high, or problems with the precision
will rise.

Sample2D is modified to be consistent with 17 MeV hypothesis.

PDF macro SigLikX17 is added.