Slight changes to GoodnessOfFit and FCgenerator
GoodnessOfFit: now the parameters for toy generators are taken from
pars, consistently with FCgenerator. The previous approach using
logL.values is just commented out. It will be removed later

FCgenerator: fixed yields definition in the loop. getYields was used
instead of getSignalYields. Now the generator is tested and ready to go