Fixed 3D likelihood profiling
The center of the constraints was wrongly assigned to the best fit when
assessing the likelihood ratio on data for a given FC point. Now this is
correctly set to the value a priori and then sampled for toys as usual.

So far DataToys for sensitivitiy didn't have random constraints. Now it
is fixed.

Added the constrainMass flag in the bestfit functions for the X17
hypothesis test.

Added testHypothesis: this allows to test specific hypothesis by
sampling toys by setting the parameters to their best fit values from
the fit with fixed signal hypothesis rather than from those from the
best fit. This is important as the BB parameters change depending on the
parameters and the residuals might not be properly incorperated in the
template bias correction. This is equal to FCgenerator with the profiled
Feldman-Cousins prescription. In the future it will be incorporated
into the FCgenerator as the default prescription.